- Sync some changes in strings from libcntpr. - Move string implementation (in "header" files) into /string dir. - Copy over libcntpr-specific implementations into _nt.c files (for now, later they could do with a define). - libcntpr's is now almost fully syncronized and could be merged into the tree except for "except" directory.
svn path=/trunk/; revision=30283

+++ reactos/lib/sdk/crt/string/strset.c
@@ -13,11 +13,11 @@
 /*
  * @implemented
  */
-char* _strnset (char* szToFill, int szFill, size_t sizeMaxFill)
+char* _strnset(char* szToFill, int szFill, size_t sizeMaxFill)
 {
 	char *t = szToFill;
-	size_t i = 0;
-	while( *szToFill != 0 && i < sizeMaxFill)
+	int i = 0;
+	while (*szToFill != 0 && i < (int) sizeMaxFill)
 	{
 		*szToFill = szFill;
 		szToFill++;
@@ -30,10 +30,10 @@ char* _strnset (char* szToFill, int szFill, size_t sizeMaxFill)
 /*
  * @implemented
  */
-char* _strset (char* szToFill, int szFill)
+char* _strset(char* szToFill, int szFill)
 {
 	char *t = szToFill;
-	while( *szToFill != 0 )
+	while (*szToFill != 0)
 	{
 		*szToFill = szFill;
 		szToFill++;
