+++ reactos/lib/sdk/crt/string/strstr.c
@@ -8,14 +8,12 @@
  *              25/11/05: Added license header
  */
 
-/* Copyright (C) 1994 DJ Delorie, see COPYING.DJ for details */
 #include <precomp.h>
-
-
 /*
  * @implemented
  */
-char *strstr(const char *s, const char *find)
+char *
+strstr(const char *s, const char *find)
 {
   char c, sc;
   size_t len;
@@ -31,5 +29,5 @@ char *strstr(const char *s, const char *find)
     } while (strncmp(s, find, len) != 0);
     s--;
   }
-  return (char *)s;
+  return (char *)((size_t)s);
 }
